// typechain-types/MailBox.ts
export interface MailBoxInterface extends Interface {
  getFunction(
    nameOrSignature: "SEND_FEE" | "getName" | "send" | "usdcToken"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "MailSent"): EventFragment;

  encodeFunctionData(functionFragment: "SEND_FEE", values?: undefined): string;
  encodeFunctionData(functionFragment: "getName", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "send",
    values: [AddressLike, AddressLike, string, string]
  ): string;
  encodeFunctionData(functionFragment: "usdcToken", values?: undefined): string;

  decodeFunctionResult(functionFragment: "SEND_FEE", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getName", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "send", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "usdcToken", data: BytesLike): Result;
}

export namespace MailSentEvent {
  export type InputTuple = [
    from: AddressLike,
    to: AddressLike,
    subject: string,
    body: string
  ];
  export type OutputTuple = [
    from: string,
    to: string,
    subject: string,
    body: string
  ];
  export interface OutputObject {
    from: string;
    to: string;
    subject: string;
    body: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface MailBox extends BaseContract {
  connect(runner?: ContractRunner | null): MailBox;
  waitForDeployment(): Promise<this>;

  interface: MailBoxInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  SEND_FEE: TypedContractMethod<[], [bigint], "view">;

  getName: TypedContractMethod<[], [string], "view">;

  send: TypedContractMethod<
    [from: AddressLike, to: AddressLike, subject: string, body: string],
    [void],
    "nonpayable"
  >;

  usdcToken: TypedContractMethod<[], [string], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "SEND_FEE"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "getName"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "send"
  ): TypedContractMethod<
    [from: AddressLike, to: AddressLike, subject: string, body: string],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "usdcToken"
  ): TypedContractMethod<[], [string], "view">;

  getEvent(
    key: "MailSent"
  ): TypedContractEvent<
    MailSentEvent.InputTuple,
    MailSentEvent.OutputTuple,
    MailSentEvent.OutputObject
  >;

  filters: {
    "MailSent(address,address,string,string)": TypedContractEvent<
      MailSentEvent.InputTuple,
      MailSentEvent.OutputTuple,
      MailSentEvent.OutputObject
    >;
    MailSent: TypedContractEvent<
      MailSentEvent.InputTuple,
      MailSentEvent.OutputTuple,
      MailSentEvent.OutputObject
    >;
  };
}